Fix building error
recovery_available was removed in order to build twrp. However, when it comes to roms, this line is required.

set_target_properties added to update build.prop and set fingerprint.

namespace removed as its not needed when building roms.

vendor_init added as this is a rom build, not a twrp build.

Added missing }

Re-add Keymaster. Set to true to prevent qcom keymaster hal from building.